Move business logic out from job to service
